requirements.txt
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.11
babel==2.17.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
dill==0.4.0
dnspython==2.7.0
docutils==0.21.2
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
imagesize==1.4.1
isort==6.0.1
Jinja2==3.1.6
makefile==1.1.0
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==25.0
passlib==1.7.4
platformdirs==4.3.8
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.22
pydantic_core==2.33.2
Pygments==2.19.1
pylint==3.3.7
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
snakeviz==2.2.2
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.42
starlette==0.47.2
tomli==2.2.1
tomlkit==0.13.3
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.4.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
apscheduler==3.10.4
reportlab==4.1.0


